backend/src/modules/users/services/UpdateUserAvatarService.ts
import IStorageEvaluator from '@shared/container/evaluators/StorageEvaluator/models/IStorageEvaluator';
import AppError from '@shared/errors/AppError';
import { inject, injectable } from 'tsyringe';
import User from '../infra/typeorm/entities/User';
import IUsersRepository from '../repositories/IUserRepository';


interface IRequest {
  user_id: string;
  avatarFilename: string;
}

@injectable()
class UpdateUserAvatarService {
  constructor(
    @inject('usersRepository')
    private usersRepository: IUsersRepository,

    @inject('StorageEvaluator')
    private storageEvaluator: IStorageEvaluator,
  ){}

  public async execute({ user_id, avatarFilename }: IRequest): Promise<User> {
    const user = await this.usersRepository.findById(user_id);

    if (!user) {
      throw new AppError('Only authenticated users can change avatar', 401);
    }

    if (user.avatar) {
      await this.storageEvaluator.deleteFile(user.avatar);
    }

    const filename = await this.storageEvaluator.saveFile(avatarFilename);

    user.avatar = avatarFilename;

    await this.usersRepository.save(user);

    return user;
  }
}

export default UpdateUserAvatarService;
